+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* Add spinner for searching in list components.
 
 ## [0.9.0-beta.5] - 2017-04-11
 ### Added